ABI: Widen the cross-memory guest address to 64 bits
## Why

`host_proc_read_bytes` / `host_proc_write_bytes` take a guest process
address. That address was typed `u32`, which is a ceiling on the
kernel-to-host contract that has nothing to do with the kernel's own
pointer width: the kernel is asking about a *different* Wasm instance,
whose address space may be 64-bit.

Nothing is observably broken today, and only by accident of there
being a single caller. Every DRI/KMS call site funnels its `u64` DRM
UAPI pointer through one narrowing guard, so a wasm64 guest handing
DRI an address above 4 GiB gets a truthful `EFAULT`. But the guard is
per-call-site, not per-contract: the next caller that forgets it
truncates silently, and a truncated address is not a failure — it is a
read or write at a *different, valid-looking* location in the target
process. Kandelo runs real wasm64 guests (`programs/wasm64/`), and
`computeProcessMemoryLayout` puts no upper bound on a host-configured
`maxPages`, so the addresses exist.

Making this the kernel's general mechanism for reaching process memory
means fixing the ceiling in the contract, once, instead of trusting
every future caller to remember a narrowing check.

## What changed

`addr` becomes `u64` in the extern declarations, the `HostIO` trait,
`WasmHostIO`'s forward, the TypeScript import bodies, and the
native host's closures. `len` stays `u32`: it bounds a *kernel-side*
buffer, and the kernel's own memory is capped at 1 GiB
(`--max-memory=1073741824`, both target blocks in `.cargo/config.toml`),
so 32 bits is provably enough — and it keeps the JavaScript value a
`number` instead of a `BigInt`.

**The TypeScript side stops hardcoding pointer width 4.** `addr` now
arrives as a BigInt, so the width handed to
`checkedWasmImportMemoryRange` is 8 — and that 8 describes how the
value arrived (an i64 import parameter), not how wide the target
process's address space is. No per-process width lookup is needed,
because width is not standing in for one: the real bound is the target
process's own current buffer length, enforced against `procMem`
directly. A wasm32 process's buffer never exceeds 4 GiB, so an
out-of-range address is rejected by the buffer it misses rather than by
an assumed pointer width. That is a truer check than the old one and
it adds no callback to `WasmPosixKernel`.

**The DRI ordering guarantee is preserved, not dropped.**
`DRM_IOCTL_MODE_GETRESOURCES` and `MODE_GETCONNECTOR` each write
several independently-addressed nested buffers, and a bad third
address must leave the first two untouched. That guarantee used to
come for free from the narrowing check, which could fail before any
write. With nothing left to narrow, `probe_dri_process_pointer`
replaces it: a pre-pass that probes each address with a
`proc_read_bytes` of the exact length that will later be written.
Reading is side-effect-free, and a range inside the process's memory
now is still inside it when the writes begin — linear memory only ever
grows, and the host dispatches the syscall synchronously without
re-entering the kernel.

## ABI

No `ABI_VERSION` bump. `abi/snapshot.json` does not describe
`env.host_*` import signatures at all
(`grep -c host_open abi/snapshot.json` is 0), so this produces no
snapshot diff; and ABI 44 is unreleased and in-dev on this branch
(`main` is at 43), which `docs/abi-versioning.md` says is amended
rather than superseded. Recorded plainly because the rule does not
decide itself: this changes *pointer interpretation* in a
kernel-to-host contract, which `docs/agent-guidance/abi.md` names as a
possible bump trigger even when the snapshot is unchanged. The
snapshot check passing is therefore NOT evidence that this change is
compatible — it is evidence that the snapshot cannot see it.

## Validation

- `scripts/check-abi-version.sh` (verify): "abi: snapshot is in sync
  with sources", "abi: ABI_VERSION and snapshot are consistent", and
  `git diff --stat abi/snapshot.json` after a regenerate was empty —
  the expected finding, not a clean bill of health (see ABI above).
- `cargo test -p runtime-core -p kandelo --target aarch64-apple-darwin`:
  1763 + 3 + 4 + 6 passed, 0 failed.
- `cargo test -p host-native --target aarch64-apple-darwin`: 52
  passed, 0 failed, 4 ignored (pre-existing).
- `npx vitest run test/kernel-public-scratch.test.ts`: 70 passed,
  including a new case proving a guest address above 4 GiB is rejected
  rather than aliased onto its low 32 bits, and that the same low bits
  offered as a 32-bit address still work.

Performance was not measured. An `i64` import parameter reaches a
JavaScript host as a `BigInt`; today only DRI pays that, and this
change puts nothing new on the syscall hot path, so no benchmark claim
is made either way.

+++ crates/host-native/src/guest.rs
@@ -1856,7 +1856,21 @@ fn define_kernel_host_imports(
     // host_proc_read_bytes(pid, addr, dst_ptr, len) -> i32 and
     // host_proc_write_bytes(pid, addr, src_ptr, len) -> i32: the kernel's
     // general cross-memory primitive. `addr` is an address in the GUEST
-    // process `pid`; `dst_ptr`/`src_ptr` are addresses in KERNEL memory. Until
+    // process `pid`; `dst_ptr`/`src_ptr` are addresses in KERNEL memory.
+    //
+    // The two address widths are deliberately different, and the closures'
+    // parameter types say so. `addr` is `u64` because it names a location in a
+    // guest whose width the kernel does not control — one signature covers a
+    // wasm32 and a wasm64 guest, and an address above 4 GiB from a wasm64
+    // guest must be rejectable rather than silently aliased down to its low 32
+    // bits. `dst_ptr`/`src_ptr` are `u32` because they name a location in the
+    // kernel's OWN linear memory, and this host runs the wasm32 kernel build
+    // (`target/wasm32-unknown-unknown/release/kandelo_kernel.wasm`, see
+    // `crate::EXPECTED_HOST_IMPORT_COUNT`'s neighbours in `lib.rs`), where a
+    // kernel pointer is an `i32`. A wasm64 kernel would import these with an
+    // `i64` in that position; matching the module's declared type is what
+    // `Linker::func_wrap` checks, so that build would need its own closure
+    // rather than silently mismatching here. Until
     // now both fell to `define_unknown_imports_as_traps`, so any kernel path
     // reaching for process memory killed a native run — truthful, but it meant
     // the native host could not run the DRI/KMS paths that have used this
@@ -1891,12 +1905,12 @@ fn define_kernel_host_imports(
         linker.func_wrap(
             "env",
             "host_proc_read_bytes",
-            move |_c: Caller<'_, ()>, pid: i32, addr: u32, dst_ptr: u32, len: u32| -> i32 {
+            move |_c: Caller<'_, ()>, pid: i32, addr: u64, dst_ptr: u32, len: u32| -> i32 {
                 if pid < 0 || pid as u32 != *current_pid.lock().unwrap() {
                     return -(libc_errno::ESRCH);
                 }
                 let guest = current_memory.lock().unwrap().clone();
-                proc_copy_in(&guest, u64::from(addr), &kmem, u64::from(dst_ptr), len)
+                proc_copy_in(&guest, addr, &kmem, u64::from(dst_ptr), len)
             },
         )?;
     }
@@ -1907,12 +1921,12 @@ fn define_kernel_host_imports(
         linker.func_wrap(
             "env",
             "host_proc_write_bytes",
-            move |_c: Caller<'_, ()>, pid: i32, addr: u32, src_ptr: u32, len: u32| -> i32 {
+            move |_c: Caller<'_, ()>, pid: i32, addr: u64, src_ptr: u32, len: u32| -> i32 {
                 if pid < 0 || pid as u32 != *current_pid.lock().unwrap() {
                     return -(libc_errno::ESRCH);
                 }
                 let guest = current_memory.lock().unwrap().clone();
-                proc_copy_out(&kmem, u64::from(src_ptr), &guest, u64::from(addr), len)
+                proc_copy_out(&kmem, u64::from(src_ptr), &guest, addr, len)
             },
         )?;
     }
